Fix "No search results" briefly showing up during the app web UI loading
Previously, the banner with text "No search results in this view for
{{ queryString }}" was shown briefly in the middle of the screen when
(re)loading the Music app web UI. Apparently, this happened because the
Angularjs framework had to be loaded before it could start dynamically
hiding things, including this banner.

Now this glitch is prevented by setting CSS rules which hide this and
other banners by default. The rule is lifted once the MainController
initialization reaches its end, by adding the class `loaded` on the
element `#app`.