Enable running unit tests without host cloud codes
Previously, the autoloader of the unit tests digged any cloud core
classes from the directory tree above the project root, i.e. it was
assumed that the music app root folder was located under apps folder of
a valid ownCloud or Nextcloud installation. With the revised
autoloader, all those core API classes are loaded either from the
composer-installed mock API vendor/christophwurst/nextcloud or from the
stubs folder. This is the same strategy as used with phpstan.

Also, the autoloader.php was direly in need of refactoring, and it is
now much more DRY.