Remove python from codeql analysis
The project has only 25 lines of self-written python code (in a build script)
but more than 100k lines of python code in total in all dependencies.
Analyzing the python code on github took around 4 minutes which is not worth
it, considering that all the python code is related to just development time
activities.