Hide a few more UI elements while AngularJS is not yet operational
This is a follow up for the commit 10553e5 which hid the "no search
results" and "no content" banners for the short moment before AngularJS
is operational and can dynamically show or hide the components. There
were still a few bit less visible components which were briefly shown
without interpolation during the page load.