Update webpack from v4 to v5
Webpack updated from 4.46.0 to 5.58.1. The motivation for this was the
security alerts on npm on the v4.

This required slight code changes, as the importing of an svg now gives
the absolute URL instead of the previous relative URL. This basically
makes it a bit simpler to use those packaged svg files.