Update Subsonic API to version 1.16.1
The only difference between 1.16.0 and 1.16.1 is the property
`artistImageUrl` which has been added to artist responses. It now holds
an URL which can be used to fetch the artist image in the original size
(as opposed to using `getCoverArt` normally, which returns a scaled and
cropped version of the image). Also the `largeImageUrl` included in the
response of `getArtistInfo` now refers to the unscaled and uncropped
image.