Remove some testing/analysis-related files from the release packages
On the other hand, the composer.json and composer.lock files are now
included in the source packages (automatically generated by github) and
removed only from the release packages. That is, they are not ignored
yet by .gitattributes but later by release.sh. The principle is now
such that only files strictly related to the CI integration are ripped
from the source packages by .gitattributes.