Fix a major performance problem in AlbumMapper::findAllByArtist
This function became horribly slow when the artist in question had many
albums, at least on the test DB which also contained lots of tracks on
another user. For artists having less than 10 albums, the performance
was still somewhat acceptable, taking a few seconds at most. But with
more than 10 albums, the time used grew rapidly, and it could take e.g.
15 seconds. This was especially problematic since some clients were
already hitting a time-out by the time the results were ready.

Apparently, the problem happened because the DB couldn't use the
defined indexes efficiently to carry out the query. But restructuring
the query a bit made wonders: the time taken to execute an example
query was reduced from ~15s to ~50ms. All this was tested on MySQL and
it is not certain that the problem or the fix is applicable to any other
DBMS.