Subsonic: Support also track/album/folder IDs in `getArtistInfo`
According to the Subsonic API documentation, it should
be possible to call `getArtistInfo` also with an ID pointing to an
album or a track. On the other hand, `getArtistInfo2` is not specified
to have such support, but we provide it anyway, because we use a shared
implementation for both of these functions (the real Subsonic server
couldn't do this because it doesn't use ID prefixes, making some of the
different ID types indistinguishable from each other).

Also, the the function now supports fetching the artist info with a
folder ID. This may be relevant when browsing the library by folders.
In this case, the info is fetched based on the folder name; if there is
an artist in the DB with the same name as the folder, then the info can
be fetched.

refs https://github.com/owncloud/music/issues/906