Stop button shown with long press on the play/pause button
Long press on the play/pause button in the controls pane now shows a context
menu with the option "Stop". This is especially to enable stopping on
touch-only devices which have no HW media control keys or the shift button to
use keyboard shortcut.

On touch devices supporting it, there is also a small haptic feedback given
when the context menu opens. It helps to notice that the menu has opened as
the menu itself is often hidden by the user's finger. This seems to currently
work at least on Chrome for Android but not on Firefox for Android.

refs https://github.com/owncloud/music/issues/911