Stop button shown instead of the play/pause while shift held down
This is a complementary method for stopping, in addition to the long-press
context menu added in 4b57695. This of course is possible only on
keyboard-enabled devices while the context menu is geared towards touch
devices. This shift+click is no easier than the previously available keyboard
shortcut shift+space, but should be a bit easier to find and to remember.