Add a third "match mode" to BusinessLayer::findAllByName
Previously, there were two "match modes" using the parameter `fuzzy`:
- fuzzy=true, where the search pattern was interpreted as a substring or
  set of substrings, and wildcards % and _ could be used
- fuzzy=false, where the pattern had to match the whole name column and
  no wildcards were supported

Now this boolean parameter was replaced with an enum-like parameter.
The previous modes correspond the values `MatchMode::Substring` and
`MatchMode::Exact`. In addition, there is a new alternative
`MatchMode::Wildcards` where the whole pattern needs to be matched but
wildcards are still allowed.

Also, the `fuzzy` parameter was removed from
TrackMapper::findAllByNameAndArtistName because this wasn't nowadays
used for anything, and it unnecessarily complicated the code.