Allow replacing forbidden characters on artist image name with '_'
It was previously impossible to set artist image for an artist having
slash '/' in its name. Also characters <>:"\|?* could be problematic
because these are forbidden in file names on Windows. Any of these
characters may now be replaced with an underscore '_' in the image file
name which is matched to the artist.

refs https://github.com/owncloud/music/issues/913