Fix "shift stop mode" added in 7ec0f6f0 getting stuck sometimes
There was such a problem at least on Firefox that, after restarting the
browser and restoring previous tabs, it sometimes had the
"shiftHeldDown" property set right from the beginning, until when the
shift key was actually pressed and released. Hence, the play/pause
button was replaced by the stop button immediately upon starting the
playback. The state variable is now reset in more places in attempt to
prevent it from getting stuck, at least for long time.