Enable providing artist image also for the "Unknown artist"
It was not possible to give artist image for the "Unknown artist" even
though the front-end hinted in the tooltip that it should be possible.

As an extra challenge, this string is localized. The image name is now
compared to the localized version of the string (as that's what the
front-end was hinting). This isn't quite trivial, because there's no
"current language" available in the context where the images are
matched to artists (within FileHook or `occ` command). Hence, we need
to manually dig out the language preference of the user being handled.