Fix Albums compact layout not scaling well in single-column mode
When the window is so narrow that only one column of albums fit, the
only column is supposed to scale to fill the whole window. This,
however, didn't work when the Albums view was using the compact layout.
Instead, the column was still enforced to the same width as when
multiple columns are shown.

This was especially prominent on landscape mobile screen where album
names got unnecessarily truncated and long track names got split to
multiple rows. At the same time, there was lot of empty screen space
on the right side of the albums.