Fix Nextcloud dark theme not always properly applied
It happened quite often with the Nextcloud dark theme that the Music
app didn't notice the theme being active, and hence, the proper styling
was not applied. This made many icons nigh invisible. This happened
especially often when the browser page was being reloaded, and a new
reload or two were needed to fix the situation.

Ultimately, this was a timing issue, and caused because the theming
information happened to not be available yet when the Music app checked
the theme. This was now fixed with a bit hacky solution: If the theming
info is not yet available upon the Music app initialization, the app
asks again once per second, ten times at maximum before giving up.