Fix scanning via the web UI often not finding the artist images
There was a stupid mistake in ApiController::scan with the
short-circuit evaluation. Consequently, if the scanning found any album
cover images, then it didn't even try to find the artist cover images.
This was especially bad because the artist images are not searched in
the background task, either (because this can be done only for one
user at a time). And especially stupid, because the front-end doesn't
even need the information here if any artist images were found or not.

The same problem was not present when scanning from the command line
with `occ`.