Never use the library root folder name as an album or an artist name
If the artist or album name is not found from the file tags, the Music
app uses the parent folder name as album name and second level parent
folder as artist name. This logic now makes an exception on the root
folder of the music library so that its name is never used as fallback
for album or artist name. It's unlikely, that the user would have named
the library root with the name of an artist or album, and if there's no
other name available, then the "Unknown artist" or "Unknown album" is a
better guess.

This also enabled a few small simplifications on the Scanner class
interface.