Fix layout problems on Nextcloud 22.2.1 and later
The problems were most notable on the albums view using the "normal
layout" where the height of tracks shown by default exceeded the height
of the album art. This caused the total height of the album elements to
vary according the number of tracks on the album, and this in turn
caused unexpected empty "album slots" in the album flow.

The root cause of the problem was the new min-height set by NC on all
button elements. This indirectly increased the line height on the track
list and hence, the total height of the track list.

Another problem caused by the same root cause was that the confirm and
cancel buttons, shown e.g. when creating new playlist, were not round
but oval.

refs https://github.com/owncloud/music/issues/923