Fix a layout issue on album names in the mobile layout
Long album names were overlapping the alphabet navigation on the mobile
layout. Fixing this was tricky because setting the percent-based
max-width didn't work in the "compact" mode where the <h2> element
doesn't have (and can't easily have) the "position: absolute"
definition. There we can use viewport-width based calculation (with the
unit "vw"), but that doesn't solve the issue in case we have the
navigation pane and sidebar visible and the app-view is much narrower
than the viewport. This final case was left unsolved.