Fix clicking on track in the Folders view when nothing is yet playing
If the user tried to start playing an individual song from the Folders
view while there wasn't already something playing, the action didn't
succeed. Instead, there was just a null reference error printed to the
console.

Fixed similar bug also from the Podcasts view. However, here it
didn't cause any visible error, but just an error being printed to the
console when the play scope was a channel and the end of the channel
was reached.