Add support for URL arguments: autoplay, shuffle, repeat
When the arg `atuoplay` is given without a value or with any truthy
value, the app starts to play the targeted view immediately when the
page has been loaded.

When the arg `shuffle` is given without a value or with a truthy value,
the shuffle feature is enabled upon page load. Similarly, giving it
with a falsy value disables the feature.

When the arg `repeat` is given with value 'one' the repeat-one mode is
activated upon page load. With any other truthy value, the repeat-all
mode is activated and with falsy values it is deactivated.

In all the cases described above, "falsy" values are 'false', 'no',
'off', '0', 'undefined', and 'null'. Everything else is "truthy".

refs https://github.com/owncloud/music/issues/922