Fix metadata not shown in the Files player for files outside the library
The lite player was showing just the file name based title and no album
art for the files which reside outside of the user's music library path.
This has been broken since the commit fa647692 released in v1.3.0.
There, the signature of the function CoverHelper::scaleDownAndCrop
changed a bit, but this one call was not updated.