Send args of all HTTP PUT requests within the request body
Previously, the arguments of the HTTP PUT requests were transfered as
URL query arguments. This was bad because ususally the URL maximum
length has been configured on the server as just a few thousand
characters. For this reason, sorting a very long playlist or removing
its duplicates was not properly saved on the server.