Fix handling of episode with identical GUID on multiple podcast channels
On the DB level, the unique constraint ensures that the triplet
user_id/guid_hash/channel_id is unique. On the other hand, in
PodcastEpisodeMapper it was assumed that the pair user_id/guid_hash would be
unique but this is not necessarily true. Hence, having the same episode in
multiple channels of the same user caused issues.

refs https://github.com/owncloud/music/issues/937