UI option to set the playback rate
A new item "Playback rate" was added to the playback context menu opened
by long-pressing the play/pause button on the controls pane. This
allows setting the playback speed between 0.5x - 3.0x compared to the
default rate. Tapping on the item title changes the rate 0.5 -> 1.0 ->
1.5 ->2.0 -> 2.5 -> 3.0 -> 0.5 but it's also possible to tap/drag the
range input element shown below the title to set the rate in 0.1x steps.

refs https://github.com/owncloud/music/issues/972