Fix error handling when playing an HLS radio stream fails
When starting to play an HLS radio stream failed, the player user to just
play the previously active radio stream without giving any errors, and still
showing the title of the failing stream on the controls pane. This was very
misleading and confusing.

Now the failure to start an HLS stream is shown in a notification and the
playback is stopped.