Make the playback within the Music app (almost) gapless
We now have two instances of the PlayerWrapper. When the playing instance has
buffered the whole song, the other instance is ordered to pre-buffer the next
song in the play queue. Then, when the playing song changes, the instances
are swapped.

This removes most of the gap when the playing song changes. What is left is
something like 100 ms of silence between the songs.