Handle failure to load the '7.html' and prevent the 'error' level logging
By default, file_get_contents emits a warning on failure to find the
requested page, and Nextcloud logs that as error (level 3). This could cause
lot of errors (one every 32 seconds) to be logged on some radio channels
which do not provide any metadata. We now tell file_get_contents to ignore
the error and log it on our own, on the 'debug' level (0).