Don't log on the 'error' level when failing to load a podcast URL
We already have a 'warning' level (2) logging in place when updating a
podcast channel fails because of invalid/unavailable URL. Hence, it makes no
sense to let the PHP warning from file_get_contents to be logged on the
'error' level (3), as would happen if we don't suppress this. As a rule of
thumb, invalid configuration on user's behalf should usually not be
considered an "error", those should be reserved for cases where the source of
the problem is in the software itself or in configuration made by the admin.