Improve the robustness of radio stream metadata parsing
On plain HTTP streams it often (but not always) happens that the first
`fread($fp, 1024)` call on the stream doesn't read only the headers but
also the beginning of the body. Apparently, this never happens with
HTTPS. This is now checked and taken into account when seeking to the
beginning of the first metadata block.

The skipping of the stream bytes before the metadata block is now made
with `fseek($fp, $bytesToSkip, SEEK_CUR)` instead of calling `fread` in
a loop. The PHP manual states "Not all streams support seeking. For
those that do not support seeking, forward seeking from the current
position is accomplished by reading and discarding data; other forms of
seeking will fail. ". That is, on non-seekable stream like the HTTP
stream, this should do basically the same thing as the previous loop
but may be better optimized.