Make the playback actually gapless by overlapping the tracks a bit
Get rid of the previous 100-200 ms gap between the tracks by starting the
playback of the new track 0-500 ms before the previous track ends, i.e. let
the two playbacks overlap this much. We can't be more precise than this
because the rate we get the progress events may vary. But usually, the actual
overlap will be in the range 200-400 ms. This little of overlap can't be
really heard but the songs are seamlessly merged together.