Fix buffering bar flickering while the next song was being changed
The flickering happened because the buffering of the previous song continued
during the debounce time of the track change. Hence, the buffering percent
was first set to 0 by setLoading and then a new buffering event fired from the
previous track and set it back to some greater value.

Fixed by stopping previous playback immediately upon song change, instead of
just pausing it. This discards the previous audio source and stops the
buffering.