Show artist and album names from Last.fm on the track details
The album and artist names as reported by Last.fm are now shown on the
Last.fm tab of the track details. This is most useful in case the file is
incompletely tagged and these names are not already available on the General
tab. It may provide interesting information also in case the track in the
user library belongs to a compilation album in which case Last.fm may show
the album where the track was originally released.

refs https://github.com/owncloud/music/issues/995