Mark channel as updated also when the operation fails
Previously, the application didn't mark the podcast channel as
"updated" if the updating failed e.g. because the RSS feed could not be
reached. Because of this, the update was retried each time the
background task was executed which could be e.g. every 5 minutes, as
long as the operation didn't succeed. If the failure was a permanent
state of affairs, then this could generate a lot of warnings to the log
file.