Request always the same metadata type as previously got for the station
If a radio station previously provided, say, Shoutcast-v1 type of
metadata, then the periodic updates are requested using the same type
(and the other types are not tested for availability). It is unlikely
that a station would change the metadata type in the middle of a
session and always testing all types would be pointless. The cached
information about the metadata type is discarded upon page reload and
all metadata types are again considered on the next try.