Support giving (non-HLS) playlist URL as radio stream
Sometimes, the public URL of a radio stream points actually to a
playlist file which in turn contains the actual stream URL. One special
case of this is the HLS-type stream and that is already handled by the
front-end library HLS.js. But now the Music app can resolve these
indirections also in the non-HLS cases.