Set tighter timeout limits for the secondary radio metadata requests
It seems that the file_get_contents may actually take at least twice as
much time as the timeout is when the site is responding but very
slowly. With the 8s timeout, reading the Shoutcast v1 and v2 and the
Icecast metadata may then take 16 seconds each, still finishing before
the default 60 s timeout between our front-end and back-end kicks in.