Control the CSP img-src header by the setting `music.allowed_radio_src`
Add only those URLs to the Content-Security-Policy header `img-src` which
have been white-listed for streaming with the setting key
`music.allowed_radio_src`. The external image sources are needed only to show
the cover images (and possible description images) of the podcasts. Probably
in 99.9% of cases, these images come from the same source domain as the
related stream and the domain has to be white-listed anyway to be able to
listen to the stream.

By default, the `music.allowed_radio_src` covers every possible URL and this
changes nothing. But this now gives possibility for a security-concerned
admin to limit the enabled image sources if so desired.

Also, clarified it in the settings view that the `music.allowed_radio_src`
key affects also podcasts.