Add configurable option to ignore articles on artist ordering
There is now a configuration item in the Settings view to define any articles
which should be ignored when deciding the correct alphabetical order of the
artists. By default, these are: 'The', 'El', 'La', 'Los', 'Las', 'Le', 'Les'
(this list was taken from the examples of the Subsonic API documentation).

In addition to the web UI, the setting affects the Subsonic API methods
`getIndexes` and `getArtists`. However, it seems that many Subsonic clients
don't actually use the order provided by the server; this is the case at
least on DSub and Ultrasonic.