Refactor albumart directive to utilize isolated scope and data binding
Only single object now needs to be passed to the directive instead of
passing two attributes "cover" and "albumart". In the previous system,
it was not obvious that these both were related to the same directive
and watching two attributes separately seemed hacky. The directive
itself now takes care of the differences between albums and podcast
channels.