Involve artist name when deciding color for album art placeholder
Previously, the color of the album art placeholder was calculated from
the album name alone. Hence, if multiple artists had an album with the
same name, then those all got the same color. This was prevalent
especially with the album "Unknown album" which is often found under
many artists.

Now the seeding of the placeholder color generation happens by
concatenating the artist name and album name. This way, each "Unknown
album" likely gets a different color.