Fix a performance problem in the cleaunup task when PostgreSQL used
Apparently, the NOT IN operation with a sub-query is a problem for
(some versions of?) PostgreSQL. The previously used query to clean
up obsolete artists could take minutes to execute when there were tens
of thousands of artists, while the new query should takes seconds or
less. The new query seems to be slightly more performant on MySQL, too.

refs https://github.com/owncloud/music/issues/997