Fix media control keys working badly on background Firefox
On (recent versions of?) Firefox, there was such a problem that the
media control keys on the keyboard reacted with an annoying delay if
the Music tab was not on the foreground. Also, if the keys were pressed
too many times during these delays, then the whole media session could
get stuck.

Apparently, this was caused by some power saving features of Firefox.
It seems that the timeout handlers of the inactive browser tabs are
handled with somehow reduced rate. The fix was to not rely on timeout
to handle the media keys. Instead, we now call $scope.$apply within the
media key handlers, basically kicking off a new $digest loop. In this
context, there's no risk of nested $digest loops and there should be no
performance concerns since the key handlers are called only
sporadically.