Fix playback breaking on IE on the first track change or stop button
On Internet Explorer, the audio element can't survive from setting its src to
an empty string. This was done as part of the stop functionality, and
nowadays, the stop function is called every time the playing track changes.
Clearing the src is now skipped on IE; it shouldn't be all that useful there
anyway because IE doesn't support the mediaSession API.