When available, use background color definitions from the cloud core
When we use the common color constants, we don't need to define as many
exceptions for the built-in dark theme of Nextcloud. Similarly, 3rd party
themes will have easier time as every element doesn't need its own custom
rule.

Note that the used color constant `color-main-background-translucent` is not
available on any version of ownCloud. To mitigate this, the hard-coded
translucent white color is used as a fallback color.

refs https://github.com/owncloud/music/pull/1002