Subsonic: Support `getCoverArt` also for albums/artists with no cover art set
The original Subsonic server provides some image for those entities not
having an actual cover art set and the clients may expect this. However, the
`coverArt` property for those entities is still omitted; it's also marked as
optional in the Subsonic REST API specification.

The placeholder image is created with similar logic as on the web UI; this
logic has been ripped from the Nextcloud source file
lib/private/Avatar/Avatar.php.

refs https://github.com/owncloud/music/issues/1000