Subsonic: Disable the SameSiteCookie middleware
For some reason, the Nextcloud middleware defined in
https://github.com/nextcloud/server/blob/master/lib/private/AppFramework/Middleware/Security/SameSiteCookieMiddleware.php
causes perpetual 302 redirection loops on some Subsonic clients in some
situations. This middleware has now been disabled on our Subsonic API.

The problem has been observed on the Android client Substreamer when loading
the album art and on the HTML5 client Jamstash. On Jamstash, the problem
occurs when using the client from jamstash.com but not when it has been
installed on the same server with Nextcloud. It's possible that this has also
something to do with the HTTP server settings used.

refs https://github.com/owncloud/music/issues/1000