Subsonic: `getScanState` in json mode to return "false" as bool
Previously, the stubbed method `getScanState` returned the property
`scanStatus` as a string "false" while the real Subsonic server returns here
the boolean value false or true. This caused the Substreamer client to never
stop polling the method since scanning was started. In XML everything is just
string anyway so there was no difference.

Also, added the number of scanned tacks in the library to the result of
`getScanState` because this is what the real Subsonic server does, too, also
when no scanning is ongoing.

refs https://github.com/owncloud/music/issues/1000