Disable the CSRF prevention on a few more REST API endpoints
The CSRF prevention was already disabled on the most REST API endpoints
but there were a few where it was still active. This made it more
tricky for external clients to use these API endpoints. On the other
hand, there was no real risk for a malicious user to cause real harm by
making a CSRF request to any of these API endpoints, apart from small
time vandalism by altering the user settings.

The CSRF prevention was left active only on the method
`/api/settings/userkey/add` as it allows granting access to the user
content using a password defined by the request.

refs https://github.com/owncloud/music/issues/1012