Add keyboard.js polyfill to fix the the keyboard shortcuts on IE
On Internet Explorer, there's natively no property `code` on the
keyboard events. The polyfill
https://github.com/inexorabletash/polyfill/blob/master/keyboard.js was
added to overcome this. Note that the `code` filled by the polyfill
depends (at least sometimes) on the used keyboard layout while the
`code` filled natively by modern browsers does not. However, this
shouldn't cause any problems on those keys which we have currently
assigned for the shortcuts.