Fix alphabet navigation breaking down with characters greater than U+FFFF
Unicode characters greater than U+FFFF (i.e. characters outside the basic
multilingual plane BMP) are encoded using surrogate pairs in javascript. Our
alphabet navigation pane didn't handle names starting with such characters
properly: The code tried to extract the first character of the name with
`String.slice(0,1)`, but in reality, this extracted only the first half of the
surrogate pair. This, of course, was not a valid character and the rest of the
logic no longer worked as intended.

refs https://github.com/owncloud/music/issues/1021