Respect the global keyboard shortcut disable switch introduced by NC25
On Nextcloud 25, the Accessibility settings contains a switch titled
"Disable all keyboard shortcuts". According to its description, it
disables all keyboard shortcuts in all applications in order to
prevent any interference with accessibility tools. The server core
can't actually enforce this but all applications are supposed to read
this setting and disable their own shortcuts when the switch is on.