Fix handling of seek preview 0:00 in the embedded Files player
In practice, noticing any difference is possible only on touch devices because
there, the preview is shown also when the touch point is to the left from the
whole seekbar. On non-touch devices, the user can't tell if the mouse pointer
is on the leftmost pixel of the bar or one pixel beyond the edge of the
seekbar, in which case hiding the preview would be the right thing to do.