Adjust to dark theme also when the theme comes from browser preference
In NC25, it's possible to select the theme "System default theme" which
may be either dark or light, depending on the browser and OS settings.
Music app now adjusts its icons and other elements properly also in
this case. Furthermore, the Music app can now handle dynamically
changes in the system theme, without needing to reload the page.