Ignore the last remaining warning from PhpStan
The function OCP\Files\File::fopen has been documented to return non-null
value in the phpdoc of the composer module "christophwurst/nextcloud" which we
use to stub the cloud core interfaces for PhpStan. However, the function may
return null values on errors at least on some of the supported host clouds.