More secure generation of the Ampache/Subsonic API keys
The API keys used with Subsonic and Ampache APIs used to be generated on the
front-end using cryptographically weak pseudo-random method. Meanwhile, there
was already available a REST API endpoint `api/settings/userkey/generate` which
could be used to generate the API key on the back-end using cryptographically
secure randomness.

The front-end was now modified to utilize the afore mentioned backend API.
Also, the API endpoint `api/settings/userkey/add` was removed altogether as
redundant and to prevent any accidental use of less secure methods. Note also
that the old method used only small alphabet and numbers in the generated key
while the new method uses those plus capital alphabet.

The `api/settings/userkey/generate` was modified to allow omitting the
description as that's how the removed API worked, too, and there's no
compelling reason to enforce the user to give the description.