Fix a minor layout issue on the play/pause context menu
When the play/pause context menu is visible and the shift is held down,
the Stop item in the menu is replaced with a Play or Pause item. This
replacement item used to have slightly different spacing than the other
menu items. This was because it had a newline in HTML between the icon
span and the text span and this got translated to a single space
character when rendering. Meanwhile, the other menu items defined both
the spans on one row.

Now all the menu items are defined with one span per row, leading to
similar spacing on them all.