Remove the obsolete background color from the #updateData element
This is the update button shown in the lower left corner while scanning
and the web UI doesn't reflect the latest state of the library. In the
current UI design, whenever this may be shown, the "Scanning" banner
is collapsed to the bottom of the screen and the update button will be
on it. Hence, the button doesn't need its own background. In the past,
the background was necessary when the "Scanning" banner was in the
middle of the screen also in this case.