Adjust icons to the dark them by utilizing var(--background-invert-if-dark)
The css variable defined by NC25 is now used to invert the icon colors when
necessary, instead of using a custom css class for the dark theme. This fixes
the dark theme handling on the link-shared folders on NC25.

For now, this breaks the dark theme handling for all Nextcloud versions older
than 25. The plan is to fix that next by implementing a kind of "polyfill".