"Polyfill" the var(--background-invert-if-dark) on older NC versions
Fix the dark theme icon color inverting on on NC < 25 after the previous
commit. We now set the CSS variable --background-invert-if-dark on our own if
it has not been set by the cloud core.

In this legacy support, we now identify the dark theme by analyzing the
background color directly. It has been a longstanding issue that the API
previously used to detect the selected theme often returned wrong values after
changing the theme, until a forced reload of the web page was conducted.