Lite player: Fix progress display when duration is not known
On the sample .aac file used, the Aurora aac.js module does not report
any kind of song duration. In this case, the lite player used to show
the song length of the previously played song. In addition, the
progress bar was frozen to the position where it was before the .aac
file was selected for playing.