Fix Firefox on Ubuntu not using multi-column layout after page load
For some reason, Firefox (107.0) on Ubuntu fires the resize events a
bit differently than the same Firefox version on Windows 10. When the
event fires on Ubuntu, the width of the #albums element is still 0. On
Widows, there are actually two events fired: on the first event, the
#albums width is 0 but on the second one it has proper width.

Worked around the problem by scheduling a new resize handling whenever
the #albums width is 0 (as that shouldn't really ever be a valid case).
This may cause some unnecessary calls but the handler function is
light-weight and this shouldn't be any kind of issue.

refs https://github.com/owncloud/music/issues/1029