Fix embedded player on single share broken after 9c96724f
The embedded player didn't work on individually shared audio files after the
commit 9c96724f. This affected only ownCloud as the Music app is not used on
those shares on Nextcloud at all. The problem was that those share pages don't
have the #app-content element but that was not null-checked.

Obviously, the dark-theme logic does not work now for the individual share page
but that shouldn't be an issue because:
a) There's no built-in dark theme on ownCloud
b) The individual share page on OC doesn't seem to be built so that it easily
    could have any non-default background color.