Try to fall back to Aurora.js if playback with HTML5 fails to begin
This is especially for cases where M4A file with ALAC encoding is attempted to
be played. From the surface, we cannot know that the file cannot be played
with HTML5 (on non-Apple browsers) as it looks like any other M4A file. It
just fails once we actually try to play it, and at that point, falling back to
Aurora.js saves the day, to some degree.