Add `occ` command `music:register-mime-types`
With the recent changes, the Music app supports playing several audio file
types which are not mapped to any MIME type by default on ownCloud or
Nextcloud. These mappings need to be added to utilize the support in the Music
app, and it can now be done semi-automatically by using the command
`occ music:register-mime-types`.