Avoid bootstrapping the embedded player when not necessary
Previously, the bootstrapping code of the Music app used to do some
unnecessary processing on many HTTP requests where it wasn't really
needed. This was because the regular expressions identifying the Music,
Files, and Share app URLs were too permissive. Now the regex's are more
focused and executed so that short circuit evaluation will be utilized.

The issue https://github.com/owncloud/music/issues/1026 might be
related to this but I haven't been able to reproduce the problem.