Fix playback jumping off the radio stream with seek forward
When seeking within a radio stream beyond the end of the buffered
stream content (using the keyboard shortcuts), the Music app used to
jump to the next radio stream. The behavior made sense when playing
files but not on live stream. On the other hand, the jumping logic
built in to the seeking function was totally unnecessary also in the
case of file playback since the jumping to the next track happens
anyway immediately after reaching the end of the track.