Fix Ampache/Subsonic key creation not working on NC <25 and OC10.0
On Nextcloud versions older than 25 and on ownCloud 10.0, calling a
controller function with the @CORS attribute doesn't work for the
normal logged-in users. Such functions are supposed to be called only
from external sites, and the user credential should be passed on each
call. The REST API `/api/settings/userkey/generate` was originally
created to be used from external sites in the PR
https://github.com/owncloud/music/pull/485, and hence has the @CORS
attribute.

This was now fixed by providing another REST API endpoint for the key
generation for our own front-end. At the same go, the REST API used by
our front-end for key management was modified to be better aligned with
the other similar REST APIs:
- URL path section `/userkey/` was replaced with `/user/keys/`
- Removing a key uses the HTTP verb `DELETE` and the target ID is part
  of the URL
- There is a getter for all the keys
- Creating the new key uses the same URL as the getter but verb `POST`

The CORS endpoint `/api/settings/userkey/generate` was kept as it
was in case some external site/application is actually using it.

refs https://github.com/owncloud/music/issues/1038